Dedupe CSV preview schemas and move to schemas directory
Move the CSV preview types into @schemas/csvUpload and add generic
PreviewCellSchema / PreviewRowSchema builders so the runtime schema is
the single source of truth: PreviewCell is now derived from the cell
builder, and the parse-csv hook reuses the builders instead of
redefining them.
