+++ src/components/CsvUpload/ValidateCsvTable.tsx
@@ -3,7 +3,7 @@ import { type Row } from '@tanstack/react-table'
 import classNames from 'classnames'
 import { useTranslation } from 'react-i18next'
 
-import { type PreviewCell, type PreviewCsv, type PreviewRow } from '@components/CsvUpload/types'
+import { type PreviewCell, type PreviewCsv, type PreviewRow } from '@schemas/csvUpload'
 import { type NestedColumnConfig } from '@components/DataTable/columnUtils'
 import { VirtualizedDataTable } from '@components/VirtualizedDataTable/VirtualizedDataTable'
