+++ src/components/UploadTransactions/UploadTransactionsValidateCsvStep.tsx
@@ -4,6 +4,7 @@ import { RefreshCcw } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
 import { type BankTransaction } from '@internal-types/bankTransactions'
+import { type PreviewCsv } from '@schemas/csvUpload'
 import { type CustomAccountTransactionRow } from '@schemas/customAccounts'
 import { DateFormat } from '@utils/i18n/date/patterns'
 import type { CustomAccountParseCsvResponse } from '@hooks/api/businesses/[business-id]/custom-accounts/[custom-account-id]/parse-csv/useCustomAccountParseCsv'
@@ -13,7 +14,6 @@ import { Button } from '@ui/Button/Button'
 import { SubmitAction, SubmitButton } from '@ui/Button/SubmitButton'
 import { HStack, Spacer, VStack } from '@ui/Stack/Stack'
 import { Badge, BadgeVariant } from '@components/Badge/Badge'
-import { type PreviewCsv } from '@components/CsvUpload/types'
 import { ValidateCsvTable } from '@components/CsvUpload/ValidateCsvTable'
 import { Separator } from '@components/Separator/Separator'
 import { templateHeaders } from '@components/UploadTransactions/template'
